GUI - move newlineandindent behaviour to buffer from mainwindow
Previously the newlineandindent functionality was implemented at the MainWindow level because that’s where all the OSC stuff was located. OSC is necessary as this functionality uses the Ruby server to figure out the indentation. However, in recent builds this caused issues for Japanese users as the keyboard events for the Return key came in differently with different locales (and different again if you switched between them?!)

By moving the event handler to the buffer code (SonicPiScintilla) the Return key events always appear to come through as the same type - regardless of locale.

Moving this code has therefore required the buffer code to learn about OSC and for quite a bit of logic to be reshuffled. This feels like the start of a good direction and a much needed push for re-organising the hairy mess that is MainWindow.

Fixes #1495