GUI - don’t shadow oscSender
previously we were shadowing oscSender and therefore creating a local variable with the same name as the one declared in the .h file.

This means that the object used in the MainWindow::sendOSC method was uninitialised and therefore totally broken!

By not shadowing it - we initialise the private field rather than a local variable and everything is good with the world.