GUI - fix return behaviour in autocomplete lists
The return key now completes an autocomplete list again.

This was achieved by sharing the auto_indent_on_run checkbox with buffers. This should only be considered an interim solution and the design of the interaction and isolation of state between the main window and the buffers needs to be properly considered for future versions. This has just been implemented like this to make it possible to quickly release a fix for the return key issue on Japanese keyboards.