GUI - check all app server ports at boot time
Check all ports that have to be listed to across the app for availability at GUI boot time.

This duplicates the checks already added to the Ruby server. However, it’s worth keeping the checks in the Ruby server for the case where a headless server is needed.

Currently only logs availability of port. Need to link this into the error reporting system so that the app throws an error as soon as a port is detected as unavailable. Alternatively - other available ports may be searched for and used if found.