Core round vals when computing ranges
When working with floats it is very easy to end up with a value of 0.4999999 instead of 0.5. This isn’t necessarily noticeable when using the value for an amplitude or a filter - but it is when used in logic such as a < b. This kind of logic was used within the range function. We now round to 14 decimal places to reduce the amount of times errors like this occur.