Sound - re-organise scsynth groups
We now no longer use two main bucket groups - synths and fx

Instead, all fx and synths go into the synths group. Also, fx create two groups - one for the synths within the block and an outer group to contain the inner synth group and the actual synth fx itself (after the synth group in the node tree).

This ensures that the syntactic ordering of synths and FX is also mirrored in the node tree order on scsynth :-)

Still more work to be done here to tidy this logic up…