Sound - remove fx group abstraction
We now only have one TL group - the current group. This is where all new synths *and* FX are placed.

New FX and Synths are placed at the :tail of the TL group - therefore the ordering of the node tree mirrors the ordering of the syntax. Synths are just placed as synth nodes and FX are placed as FX container groups which internally contain a new group for the synths created within the FX block and a FX synth node which sits *after* the FX synth group.

For example the following code:

play 50
play 70
with_fx :reverb do
  play 72
  play 75
end

would produce the nodes in the following order:

synth     synth   FX Container
s1,            s2,     [[s3, s4], fx1]