Server - detect and report failure when loading samples
Teach Server to detect when a buffer_alloc_read call to scsynth fails (typically due to the file not being one of the supported file formats such as wav, aiff or flac) and ensure the error is correctly reported back to the user as an Exception.

This is a little trickier to implement than perhaps expected due to the sample loading mechanism being asynchronous (for load speed).