GUI - inhibit log interaction when running
This stops the log from behaving weirdly when the user clicks the text area during execution. This would set the cursor to the place of the click and then all new edits would be printed at that position - not the end of the document as expected.

We now do the following:

* On initial run of code - disable user interaction with the logs and then force the cursor to the end of the document.
* On all run completion - re-enable user interaction with the logs.

Note, these two events happen at the same time as the scope start and stop actions.