Fix sample file path drag & drop
Previously, when using the drag and drop feature to output a sample's file path, on Windows this was causing an undesired '/' to be output at the front of the path.
According to https://bugreports.qt.io/browse/QTBUG-6868?focusedCommentId=214885&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-214885 the function that was used to do this, Qurl's .path(), was still behaving correctly in this instance - it is only designed to output the path component of a URI - *which may not be a valid file path on the file system*.

A comment in the above link states that Qurl's .toLocalFile() will give a valid path, so .path() has now been replaced with .toLocalFile() accordingly.