Events - prune event matchers
With this new event history design, syncers that need to wait for a matching event are added to a list of ‘event matchers’. Each incoming event is checked against all these event matchers and the ones that match will then free their respective syncers (through a Clojure-like promise system).

It is possible for this list of matchers to continually grow if many threads are continuously put into in a state of syncing by waiting for a matching event which will never arrive.

We now remove (prune) any unmatched matchers that were created by a given thread when it dies. This is made possible by storing the ID of the thread that created the event matcher in the matcher object itself. Then on death, threads will now prune the matcher list for events that have the same thread ID.