History - teach sync to handle /foo/** matchers
As the sync matchers use standard regexps and have a different code path to get matchers it’s possible that they still have a slight difference in behaviour. Here’s one of those examples - where a glob ** at the *end* of the matcher was handled incorrectly by the sync matchers but correctly by the get matchers. This is now resolved.