CueEvent - add BPM to event info and remove metadata
I had considered to use the metadata to add in optional data such as the BPM and beat values (in addition to other things I hadn’t yet thought of). However, it seems more important to be able to encode an event as an OSC message which would be more tricky with two data structures of variable length - the arg list and the metadata. One data structure of unknown length is easy if it’s at the tail of the OSC args, which is why the BPM and beat values have hard-coded positions at the start of the event arguments.