Core - teach stop to tidy up thread
Without this patch, a call to stop will render the current thread inert - meaning that if it is a named thread, further similarly named threads couldn’t be created. This was observed  when using stop within a live loop, and then after commenting out stop to #stop and re-running the code, the live loop wouldn’t spring back into action.