Sound - add live_audio
Streams audio from sound card to current sound context. Only one live_audio for a given name may exist at once - redefining the live_audio will not create a new synth, rather it will move the existing synth to the new context.

For example:

with_fx :reverb do
  live_audio :vocals, input: 1
end

This will pipe audio through your first input on your sound card into a reverb FX.

You may use live_audio anywhere you use synth or sample. It is thread safe, so may be safely used across many live loops simultaneously (to move the audio in and out of contexts dynamically should this be the desired goal).

This has been the result of a *lot* of design work and an incredible amount of fiddly coding to make it not leak memory, and be thread safe (with very careful tracking of synth creation and destruction/movement). Huge hugs to those that get pleasure out of it, this is a gift to you :-)