Remove duplicated function definition
It seems that 'midi_poly_pressure' was accidentally defined twice. The
only difference between the two definitions was that one said 'You may
also optionally pass the pressure value as a floating point value' and
the other said 'You may also optionally pass the control value as a
floating point value'. I left in the one whose indentation had been altered
recently, which was also the one that correctly referred to 'pressure'
as above.