OSC - prefix incoming path with / if necessary
It is still possible to send OSC messages out without a prefixed / in the path - however, this is because there are likely to be some OSC servers out there that haven’t been written to the full standard and may expect no initial / in the path. We want to be able to be flexible with who we can talk to, and be strict on what incoming messages look like and are handled.