OSC - only show listening message once settings are loaded
The OSC server is (re)started twice when Sonic Pi is opened.
Once when the ruby server runtime is initialised, and again when GUI
preference settings have been loaded.

Previously, during the runtime's initialisation, the OSC server was
always started with open = false, silent = false - since the GUI
preferences had not been loaded at this stage. The first log message
about the OSC server therefore always said that the server was only
listening locally, regardless of what the GUI preference had been set to
in previous sessions.

The second time the OSC server was loaded, (as a result of the GUI
honourPrefs() fn), 'open' was correctly set to whatever the preference
had been set to in the previous session, but you are never told this in
the log, since the __restart_cue_server! fn was called with silent =
false.

This commit swaps the value of 'silent' for the two calls to
_restart_cue_server! during startup, so that the first log message about
the OSC server, (always saying listening only locally) is not displayed
- and the second OSC server log message, which will respect the GUI
preference setting about listening locally and possibly remotely, *will*
be displayed.