Time State - auto trim old events
If more than 20 events are stored for a given path, it will start dropping events created more than 30s ago.

Note, it’s still possible to arbitrarily grown the time state store by creating an infinite number of events with different paths. This needs to be addressed in the future.