FX - add missing nbpf FX
I noticed recently that somehow, the synthinfo metadata for the nbpf fx was defined, but the actual synth design & compiled scsyndef were missing.
I have now added the nbpf back into the fx designs, (with the related dot and pdf file) and added the compiled synthdef (which was tested on OSX).