Stop ls_samples from matching files twice
Previously, in the ls_samples fn, the pattern given to Dir.glob was
matching any found sample twice - once because it matched the lowercase
pattern, and again because the uppercase pattern would also be a match
for samples named in lowercase. (The pattern is not a regex, so case
sensitivity is also handled a little differently).

In order to get around this problem, sets of characters are used, eg:
[wW][aA][vV], instead of matching on either of {wav, WAV} etc.