Make the documentation build reproducibly
Whilst working on the Reproducible Builds effort [0], we noticed
that sonic-pi could not be built reproducibly.

This is because it iterates over internal data structures (hashes)
in a non-deterministic order. It also sorts a list only concerned
about the length of the fields, rather than (also) considering the
alphanumeric sort, leading to nondeterminstic output.

 [0] https://reproducible-builds.org/