Revamp nested_pattern as pattern
Nested patterns were implemented in an unloaded file since version 2.8.0 before rings were available. This commit brings them up to date to use the ring API, with the addition of a special method called `.tatum` to access the time to sleep for. The choice of `tatum` as the method name builds on a convention of using this to refer to the smallest rhythmic subdivision that a listener perceives: https://en.wikipedia.org/wiki/Tatum_(music)