GUI - fix mark selection navigation shortcuts
For a long time, Sonic Pi has supported Emacs-style mark selection manipulation:

* Hit C-SPACE to move into mark selection mode
* Use Emacs-style navigation keys (C-p, C-n, C-f, C-b, etc) to move the mark (and therefore dynamically extend, reduce the selection)

Unfortunately, more recent releases of QScintilla, the underlying text editor widget, broke support for this when attempting to move the mark in some cases.

We now circumnavigate this bug by coding our own idea of ‘selectionMode’ status and using that to move (or not move) the selection where it was previously automatic.

Old behaviour is now restored with the more recent QScintilla builds!