Runtime - don’t put version number in empty buffer comment
Previously we placed the version number in the comment string for empty buffers:

```
#Welcome to Sonic Pi v2.11.0
```

This is nice, however when users upgrade to future versions and these buffers have been untouched, they don’t get updated and report the wrong version number.

Instead, let’s not put the version number in the comment!

Fixes #1169