Fix tutorial link
I was having a look at the online copy of this tutorial, and I noticed that the link from the table of contents (left pane) to the Minecraft Pi - Basic API subsection is incorrectly numbered (11.1 instead of C.1), and so it links to the wrong section of the doc. Hopefully this should fix it?