Sound - fix arg scaling issue
Previously the code attempted to optimise the amount of args that are sent over the wire from Sonic Pi to SuperCollider. If the value to be sent is the same as the default value from the synth, then it wasn’t added to the list of values to send.

However, this caused an issue where if the new scaled value was the same as the default, it wasn’t added to the args, which meant that the original non-scaled value would remain unchanged. This has now been fixed.

Many thanks to @MineMaster284 for highlighting this issue and to @Ethancrawford for pinpointing the location of the error.

Fixes #1843