Core - No longer throw out of time warnings in real time threads
If the user has selected “Enforce Timing Guarantees” in the app preferences, Sonic Pi will not send messages it knows will be late and instead throw a warning. However, in the case of live loops running in real time mode, every message *is* late thanks to reality (we’re essentially trying to react as quickly as possible). Therefore, don’t throw timing guarantees for threads that are in real time mode :-)