OSC - breaking change - switch to : as internal separator for OSC source path segment
The main issue with the previous `/osc/*/*/foo` approach is that you can't ask the question “match the next incoming `/foo` from *any* source” if the source part of the path is an unknown amount of levels deep

Originally (prior to the previous patch) you could do: `/*/foo` but if the osc source description is 3 path-levels deep, midi is 1, and others may be arbitrary levels, you can only do: `/**/foo` which matches both `/foo/` and `/bar/foo` and there's no way of only matching the first and not both.

We now still need to special-case the matchers for the first path segment with internal `:` separators and keeping it to just one level in the path works well in all regards and then provide a toggle to optionally display the `:` metadata.

We should also consider having the shortcut /osc/foo for /osc:**/foo to retain backwards compatibility