Core - add new fn eval_file
This allows you to evaluate the contents of an external file as if it was a local function.

This means that it executes within the current thread with the current thread locals (and can also modify the current thread locals). This is distinct from run_file which will execute the external file as an entirely separate ‘run’.