Sound - Don't print messages for blank nodes
Previously, when attempting to control blank nodes (eg rest notes),
The messages that got printed to the log panel tried to print out
the node id. Since BlankNodes don't have ids set, this made the
resulting message look a little odd.

We now avoid printing messages to the log if we attempt to control
a BlankNode.