Switch to ruby-style function parameters in documentation
There are many places where lisp-style function invocation syntax is
used in the documentation. This clearly works, but the semantics are
confusing to a novice (based on watching people learn SonicPi without
knowing Ruby, and in some cases even when they DO know Ruby). This is
more pronounced when the result is then method-chained, which is
frequent.