GUI - fix run syntax highlighting
Current line is now highlighted correctly on run.
Also fixed bug which could result in a buffer stuck in the highlighted state. This was caused by the unhighlighting logic using the notion of ‘current buffer’ which could have changed due to fast buffer switching. Now only the specific buffer is unhighlighted correctly.