GUI - fix colour mode shortcut switching
The keyboard shortcut for switching between light and dark modes has not been working recently.
Some extra work was needed to allow it to handle the extra colour modes that have been added to the GUI.
This has been done, and the keyboard shortcut now correctly cycles through the available colour modes.