GUI - add new Studio mode feature flag
This GUI flag is enabled if the user is a supporter of the project and will toggle special “Studio” features.

This flag is always set to true (therefore enabling Studio mode) on Linux-based operating systems.

Currently, only Patreon supporters gain access to Studio mode. This is made possible by applying a cryptographic hash of each supporter’s email address (+ a secret salt) and downloading that anonymous list of hashes. The Sonic Pi GUI then may check locally by applying the same cryptographic hashing function over a locally stored email address + the same secret hash. If there is a matching hash in the list, then the user is a supporter. In the future other supporting systems than Patreon may be supported by working with unique strings such as email addresses. These additional systems would either need to be managed manually (adding unique strings to a special file which is merged with Patreon’s email addresses) or automatically if there is an API available.

Simply knowing (or guessing) one of the existing email addresses of a Patreon will be sufficient to break this check. It is not designed to be highly secure, rather it is designed to provide some motivation and encouragement for people to support. Providing each user with their own secret salt would be one way to make the security more robust if necessary (but would require systems for managing and communicating individual salt strings).

This patch does not include any visible GUI elements for manipulating this checking process. This is to come in a future patch.