Sound - Fix the `on:` opt when playing chords
Previously, when passing the `on:` opt in an args hash to a chord playing
function (play, synth, etc etc) this args hash was commonly being associated
with every note in the chord. However, when a synth node was created for each
note in the chord, the `on:` opt was being deleted from the args hash; meaning
that after the first note's synth was created, all further notes in the chord
didn't have any `on:` opt and thus by default were audibly triggered.

We now avoid this problem by making sure the args hash is duplicated for each
note in the chord, rather than being directly reused.

Fixes #2074