Core - fix bug - stop defonce executing multiple times
When the body of a defence has a long running piece of code, the previous behaviour was to wait until this code had completed before defining the value. We now define the value first (to an initial value) and then again once the block has completed executing, this time using the result of the execution.

This stops multiple threads which are all calling defence from executing the block of code multiple times. The first thread wins and the others do not run the code.

The code is also run in a new thread therefore not affecting the execution of the main thread.