sync: consistely sync to contemporary cues
This change alters the behavior of cue/sync to not depend on the
execution order of threads. As a result, a cue will consistently trigger
a sync (exactly once) even if they happen at the same time.

Currently cue/sync matching requires that the cue to have been emitted
later than the sync that is waiting for it. This behavior neatly avoids
undeterministic behavior due to races but also makes it complicated to
reason about whether a particular cue would trigger a sync or not.

Specifically, a cue triggering a sync depends on the thread execution
order which while following consistent rules is non-trivial to determine
in the middle of a live coding session.

Consider these two threads:

in_thread do
  sync :foo
  puts beat
end

in_thread do
  cue :foo
  sleep 0.1
  cue :foo
end

Executing in the order above, the puts will emit a 0.0. If the order of
threads is switched, the puts will emit a 0.1.

A more practical situation involves syncing up live loops that sleep the
same amount of time internally:

live_loop :click do
  sleep 1
end

live_loop :music do
  sync :click
  4.times do
    play :a4
    sleep 0.25
  end
end

Contrary to intuition, the :music loop will skip execution every other
beat. Admittedly, there are ways to work around this but they are all
awkward in one way or another.

With this change, both examples above work the same (intuitive) way
irrespective of the order they appear in the buffer and in case of
live_loop, the order in which they are written and run in the live
coding session. That is, the in_thread example will always output 0.0
and the :music loop will always trigger every beat.

See this thread for more discussion, examples and workarounds being
employed:

https://in-thread.sonic-pi.net/t/synchronization-between-sync-and-sleep/2429