GUI - don't naively load icons into memory
Previously we were loading the icons using the default QIcon constructor. Whilst the original png icons were small, they had large dimensions and these directly converted to a raw internal format with a *very* large memory footprint.

We now use the `.addFile` method on QIcon to load the file into a specified size which massively reduces this internal footprint.

On macOS, the memory footprint of the loaded app binary is now reduced from ~900MB to ~170MB. This reduction *should* also be similar on other operating systems.