GUI - ensure main editing widget is in focus at boot
On some releases it has been noted that the text area cursor wasn't visible after booting. This could be resolved by switching away to another OS app and then back again to Sonic Pi.

This issue appears to be resolved with this patch which essentially emulates the act of switching back to the app during boot time. Fingers crossed....