Scripts - teach fetch-url to only fetch periodically
Fetch now takes an extra argument check_window_size which allows you to specify the minimum duration between fetches. Previous fetch results are cached in the settings and are used until it's time to make a fresh check.

This enables the behaviour of the GUI update checking to be emulated - i.e. only check every 2 weeks.