Contributors - add the inimitable Joe Armstrong
Hello Joe. It's hard to write this.

Joe was a deep inspiration, guide and important contributor to Sonic Pi,

We had so many more plans but sadly he left us too early.

At the bare minimum Joe contributed the Erlang scheduler which drives the timing of external OSC and MIDI systems. He was also an inspiration with his insightful and inquisitive questioning of design decisions and suggestions of future directions.

The most important aspects of Sonic Pi Joe helped create are yet to come. I'm excited to work on them, but less excited not to be collaborating directly with Joe on them. Still, his questioning spirit will never leave.

Joe's code filled the Royal Albert Hall with timely synth hits. Those sounds filled the hearts of thousands watching.

Goodbye Joe.