Core - fix ramp behaviour with tick
Previously ramps didn't behave as expected with ticks:

```
      r = (ramp 1, 2, 3)
      assert_equal(1, r.tick)
      assert_equal(2, r.tick)
      assert_equal(3, r.tick)
      assert_equal(3, r.tick) #=> failed as this returned 1!?!
```

This was because `.tick` now implicitly converts the list to a ring which in the case of a range, converted it to an ordinary ring. This is now fixed and calling `.ring` on a range returns a range. Indexing with `.tick` now works as expected again :-)