Style non-native file dialogs
This updates the style of the Qt file dialogs that are used on linux OS's, since
previously the text in parts of these dialogs was almost invisible. They're now
fixed.
An issue that still remains is that in the dark theme, if a scroll bar is needed
in the file dialog, it merges into the file details panel as that has the same
colour for its background. A decision may need to be made about whether or not
to change this in some manner.

Also in this change, I've also added in 'menuBarColor' to the comments
describing the list of customisable theme components, since I missed that last
time.