Sound - Breaking change - improve sample's start/finish/onset/slice opt semantics
The `start:` and `finish:` opts for `sample` now work within the specified slice or onset. Where both the slice and onset are specified, then the slice operates within the onset. Where the slice, onset and start/finish opts are all used, the start/finish operate within the slice which in turn operates within the onset.

See updated sample example for more info.