Fix - stop incoming events modifying hash during iteration
Prior to Ruby 2.7 this didn't appear to cause an issue. This may be because the bug was masked or because it wasn't a bug with the previous implementation of Ruby's hash.

However, in Ruby 2.7 when we attempted to remove an event handler from the current set of handlers whilst working through them we get a run time error. We now simply collect the names of the handles we wish to remove during the iteration process and remove them afterwards.

This fix should work on previous versions of Ruby.