[Rakefiles] Add more arguments to server:bundle_gems task, and move macOS only gems to the :main Gemfile group
The extra arguments in the bundle_gems task allows the use of --standalone and --clean options in bundle install.
Note: the server:bundle_gems task no longer has the 'path' argument as a hard-coded value is now used for the bundle path.
The :macOS group in the Gemfile seemed redundant, as we use install_if to check if the system is macOS anyway.