GUI - save and load colour theme correctly again
We now save a string version of the style name into the settings and convert that to a style enum on load.

Perhaps need to revisit the names of things here - SonicPiTheme::Theme should probably be Style to reduce confusion with the enum describing the theme and the theme class itself.

Also feels a bit weird to store the style enum in pisettings rather than use the top level theme object directly. Unecessary duplication?