GUI - theme/style internal name refactoring
Rename the SonicPiTheme::Theme enum to SonicPiTheme::Style

theme now refers to the full theme object (which can be one of a few styles) and style to the enum representing a specific style (which style the current theme is set to).

This should hopefully make the code a little clearer to understand.