DPI Fixes Part 2
My previous attempt at using auto scaling failed because the scopes were
broken.  Digging into this, I found that only on windows, the scope was
trying to use an old-style OpenGL Canvas (on windows only).  This is
most likely not DPI friendly, and I bet there is no performance
difference to using it (the docs suggest it is for better customization,
not speed).  Using the same canvas as on Mac, everything works well.

This fixes the toolbar icons, the menu text (I removed the icons, as in
Mac), the startup splash screen, and the error dialog.

One oddness: In Dark/Pro Dark modes it is hard to see the 'close'
buttons on the doc panels.  I think this is a style choice though, not a
DPI issue.  I suspect a color: white is missing somewhere in the dark
themes.

This fix will need checking on Mac and Low DPI to ensure there are no
odd behaviors; it should be OK though.