Add a simpler Build using CMake
This is, for now, a windows only cmake build setup.  It should be
possible without much effort to move all platforms to this simpler
approach. Additionally, when the ruby side has fixes for the
workarounds, this build will be even simpler.

- Updated INSTALL-WINDOWS with instructions

- Dependencies added to external/ folder with commands to make them as
part of the project.  This removes the need to build/install Scintilla,
Qwt, Boost, osmid.  They are compiled as part of the build process.
This also makes it easier to debug/fix these components if necessary

- Added a prebuild.bat to do the translations using ruby

- Moc files for scintilla not needed now

- Fix for memory alignment in server_shm.hpp; on windows there is a
chance that the string type will be misaligned with what was built into
SuperCollider (or on any platform for that matter).

- Added a config file generated from makefile; usually useful for
injecting configure-time state into source

- git ignore fixes

- Remove the ruby-aubio-prerelease from vendor; it causes a conflict and
I do not believe it is needed on other platforms too?

- Osmid is copied into the correct place (and built).