An example of how to draw in a Qt Window.
Sam asked me about rendering into a Qt window.  This is a quick and
simple way to demo how it works.  It's just a window with an overriden
paint method.  It should be nice and fast.  With some effort and
cleanup, it would be easy to ditch Qwt and do a nice visualizer that did
various interesting things.

Note: to make this work I pull out the captured scope data, so that widget
needs to be enabled.  To properly do this, the data grabbing from
SC should be in its own class, probably on a thread.

The visualization itself is simply a stereo waveform and a blue box
which approximates amplitude; just to show a couple different ways to
draw.