Fixing Onset by building Aubio
This is currently windows only, but fixes aubio by adding external
projects:
- Aubio
- Ogg
- Vorbis
- Opus
- FLAC
- linsndfile

All these components are now built as part of the prebuild step, and the
final resulting aubio dll is copied into ruby/bin

There's a lot of CMake voodoo going on here to make these components
build.  The majority of it is in external/CMakeLists.txt

With this build of aubio supporting FLAC, it can now load a flac file
and detect the onsets.

To try it (change the number at the end):
sample :amenbreak, onset: 6