Build - rename and tidy qt build scripts
We are moving to using cmake as the primary build system.

Move the many old (and often incomplete) build script into an `old` directory. The aim is to delete this directory at some point in the future. However, these scripts may have some useful information at this stage and should only be deleted when the new approach works on all platforms and is fully documented.

Also rename the windows scripts to have a `win-` prefix and the shell scripts to have a `unix-` prefix.

We should perhaps consider creating a `scripts` directory for these. Also we might consider putting the GUI source files  and directorys `.cpp`, `.h` etc into a `src` directory.