GUI - Fix issues with fullscreen mode on Windows
When in fullscreen mode on Windows, the autocompletion and other popup subwindows were not visible, despite still being clickable.

This is because there's a known issue with OpenGL in fullscreen mode and a fix is to place a single pixel border around the whole app. This is documented here: https://doc.qt.io/qt-5/windows-issues.html#fullscreen-opengl-based-windows