GUI - ensure dpi functions don't round down to 0
As the functions work on double precision values, when they get converted to integers to be used as px units, if they are between 0 and 1 then they will be reduced to 0.

This means that thin/small things will be reduced to 0 and therefore disappear.

Now, if the user explicitly wants 0 we return that. In the case of higher numbers, we return at least 1 as the smallest scaled value.