WIP: adding autotuner effect (aka a vocoder)
As discussed in the forum here
https://in-thread.sonic-pi.net/t/prototype-of-a-vocoder-in-sonic-pi/1247/2
and in issue #1930 there's some interest in having a vocoder effect in
Sonic Pi.

I prototyped a very simple one using noise and filters but it didn't
sound very good. Then I came across an excellent implementation of a
PSOLA pitch shifter as a SuperCollider quark at https://github.com/dyfer/PitchShiftPA

This works by separating the pitched content of an input from the
formant sounds (percussive, non-pitched sounds like "ssss" and "ttttt") which,
together make up a voice.

With no args the effect works as an autotune, moving the pitch to match
the nearest midi note.

When a `target_pitch` argument is used, the input is pitch shifted to
match that pitch which gives that classic "robot singing" sound that
people associate with vocoders. This can then be changed using the
`control` method to create new melodies.

Opening as a WIP commit as it still needs documentation.