GUI - further high DPI monitor improvements
We now scale all px size values in qss and css stylesheets.

Currently this scaling method is pretty naive and uses *many* passes. It definitely could be improved and optimised.

Fonts are now unified to be one of small, medium, large, x-large and xx-large (which are converted to scaled px values).

Removed specific stylesheets for widgets - preferring using a single one for everything.